Incorporated sounds (Sound Doggo's doing)
Where have all the good men gone
And where are all the gods?
Where's the streetwise Hercules to fight the rising odds?
Isn't there a white knight upon a fiery steed?
Late at night I toss and I turn
And I dream of what I need
I need a hero
I'm holding out for a hero 'til the end of the night
He's gotta be strong
And he's gotta be fast
And he's gotta be fresh from the fight
I need a hero
I'm holding out for a hero 'til the morning light
He's gotta be sure
And it's gotta be soon
And he's gotta be larger than life!
Larger than life
Somewhere after midnight
In my wildest fantasy
Somewhere just beyond my reach
There's someone reaching back for me
Racing on the thunder and rising with the heat
It's gonna take a superman to sweep me off my feet
I need a hero
I'm holding out for a hero 'til the end of the night
He's gotta be strong
And he's gotta be fast
And he's gotta be fresh from the fight
I need a hero
I'm holding out for a hero 'til the morning light
He's gotta be sure
And it's gotta be soon
And he's gotta be larger than life
I need a hero
I'm holding out for a hero 'til the end of the night
Up where the mountains meet the heavens above
Out where the lightning splits the sea
I could swear there is someone, somewhere
Watching me
Through the wind, and the chill, and the rain
And the storm, and the flood
I can feel his approach like a fire in my blood
I need a hero
I'm holding out for a hero 'til the end of the night
He's gotta be strong and he's gotta be fast
And he's gotta be fresh from the fight
I need a hero
I'm holding out for a hero 'til the morning light
He's gotta be sure
And it's gotta be soon
And he's gotta be larger than life
I need a hero
I'm holding out for a hero 'til the end of the night
He's gotta be strong and he's gotta be fast
And he's gotta be fresh from the fight
I need a hero
I'm holding out for a hero 'til the morning light
He's gotta be sure
And it's gotta be soon
And he's gotta be larger than life
I need a hero
I'm holding out for a hero 'til the end of the night